+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Master
-* 
+* add option `--v1_order` to use the old class order from v1
 
 
 ## Release 2.0.0